refactor: reorganize Zalo webhook dependencies and remove unused signature verification code
- Deleted the deprecated `deps.py` file containing Zalo webhook signature verification logic.
- Updated import paths in `zalo_webhook.py` and test files to reflect the new structure under `zalo_webhook_deps`.
- Ensured that the codebase is cleaner and more maintainable by removing unused components.
